reorganize integration tests
