+++ lib-xps/tests/integration.rs
@@ -3,4 +3,5 @@ mod integration_util;
 mod it {
     mod contact_ops;
     mod messaging;
+    mod xps;
 }
